requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.6
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
dateparser==1.1.0
DateTime==4.3
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.29.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
ipykernel==6.9.1
ipython==8.1.0
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
Mako==1.1.6
mariadb==1.0.9
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.1
mysql-connector==2.2.9
nest-asyncio==1.5.4
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pangres==4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.28
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.7
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.2.0
TA-Lib @ file:///D:/userData/SynologyDrive/Projekte/Python/Projekt_Python_Binance/TA_Lib-0.4.24-cp310-cp310-win_amd64.whl
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
wcwidth==0.2.5
websockets==9.1
yarl==1.7.2
zope.interface==5.4.0
